intepreter: skip ahead to next interrupt, when halted
This may ahead too much ahead, exceeding too much the timeout set  in
the main loop. I set a limit of 1 frame, so it will not skip a entire
frame at least, but it may still block the main loop from processing
input, increasing input delay.

I the above is theorical, didn't test any of that.

+++ core/src/gameboy.rs
@@ -334,8 +334,8 @@ impl GameBoy {
     }
 
     /// Advance the clock by 'count' cycles
-    pub fn tick(&mut self, count: u8) {
-        self.clock_count += count as u64;
+    pub fn tick(&mut self, count: u64) {
+        self.clock_count += count;
     }
 
     pub fn update_next_interrupt(&self) {
